H1_low_level_isaac_sim/requirements.txt
# Core Python packages
numpy
torch
matplotlib
scipy
open3d

# Computer Vision and Image Processing
opencv-python
pillow

# mpc
pypxr
casadi
acados_template

